ai_scribe: bump up the number of retries
The vast majority (all?) of the aborts in script generation now come from empty FC blocks.

They're in vanilla scripts, and so we'd have to remove them. We're also susceptible to them if they enter through contraction in some way. In practice, and given the typical command graph, 100 seems to be an unreachable number of aborts, so as to push the probability of their occurrence to near zero, but still possible.