command_graph.py: fix misidentification of command bytes
Using pure lookback in the script, it was "detecting" an orphaned target command byte, but this was actually the arguments from a previous (not target) command, and as such a false positive.

This was partially (wholly) responsible for infinite loops as the main block loop was trying to force the block to close, but it mistakenly though it was orphaning the misidentified target command byte, and was resetting itself in a closed loop.